+++ CHANGELOG.rst
@@ -6,10 +6,24 @@ dcicwrangling
 Change Log
 ----------
 
+1.1.0
+=====
+
+`PR 90: Geo2fdn update <https://github.com/4dn-dcic/dcicwrangling/pull/90>`_
+
+* Uses openpyxl in geo2fdn script - now this script handles ``xlsx`` files.
+  This is compatible with files produced by Submit4dn v2.2.0 and up.
+  ``xls`` files are not supported anymore.
+* Remove xlrd, xlwt, xlutils dependencies in the entire repo, as this script was
+  the only remaining one to be upgraded to use openpyxl.
+  This could be a breaking change in case of untracked scripts/notebooks.
+* Bug fix in geo2fdn script: FileFastq tab is now written properly.
+
+
 1.0.0
 =====
 
-`PR 89: update dependencies <https://github.com/4dn-dcic/dcicwrangling/pull/89>`
+`PR 89: update dependencies <https://github.com/4dn-dcic/dcicwrangling/pull/89>`_
 
 * update some dependencies including using dcicutils 4.1 and up - this causes some breaking changes and it is likely that some notebooks may stop working - especially copies in untracked folders
 * updated script utils functions for validation and the scripts that use them to allow specification of key and keyfile as strings rather than previously used key dict passed to key argument